+++ photutils/psf/models.py
@@ -456,8 +456,8 @@ class FittableImageModel(Fittable2DModel):
             input indices will be multiplied by this factor.
         """
         if use_oversampling:
-            xi = self._oversampling[0] * (np.asarray(x) - x_0)
-            yi = self._oversampling[1] * (np.asarray(y) - y_0)
+            xi = self._oversampling[1] * (np.asarray(x) - x_0)
+            yi = self._oversampling[0] * (np.asarray(y) - y_0)
         else:
             xi = np.asarray(x) - x_0
             yi = np.asarray(y) - y_0
@@ -505,19 +505,13 @@ class EPSFModel(FittableImageModel):
     norm_radius : float, optional
         The radius inside which the ePSF is normalized by the sum over
         undersampled integer pixel values inside a circular aperture.
-
-    shift_val : float, optional
-        The fractional undersampled pixel amount (equivalent to an
-        integer oversampled pixel value) at which to evaluate the
-        asymmetric ePSF centroid corrections.
     """
 
     def __init__(self, data, flux=1.0, x_0=0.0, y_0=0.0, normalize=True,
                  normalization_correction=1.0, origin=None, oversampling=1,
-                 fill_value=0.0, norm_radius=5.5, shift_val=0.5, **kwargs):
+                 fill_value=0.0, norm_radius=5.5, **kwargs):
 
         self._norm_radius = norm_radius
-        self._shift_val = shift_val
 
         super().__init__(data=data, flux=flux, x_0=x_0, y_0=y_0,
                          normalize=normalize,
@@ -586,8 +580,8 @@ class EPSFModel(FittableImageModel):
     @FittableImageModel.origin.setter
     def origin(self, origin):
         if origin is None:
-            self._x_origin = (self._nx - 1) / 2.0 / self.oversampling[0]
-            self._y_origin = (self._ny - 1) / 2.0 / self.oversampling[1]
+            self._x_origin = (self._nx - 1) / 2.0 / self.oversampling[1]
+            self._y_origin = (self._ny - 1) / 2.0 / self.oversampling[0]
         elif (hasattr(origin, '__iter__') and len(origin) == 2):
             self._x_origin, self._y_origin = origin
         else:
@@ -656,8 +650,8 @@ class EPSFModel(FittableImageModel):
 
         # Interpolator must be set to interpolate on the undersampled
         # pixel grid, going from 0 to len(undersampled_grid)
-        x = np.arange(self._nx, dtype=float) / self.oversampling[0]
-        y = np.arange(self._ny, dtype=float) / self.oversampling[1]
+        x = np.arange(self._nx, dtype=float) / self.oversampling[1]
+        y = np.arange(self._ny, dtype=float) / self.oversampling[0]
         self.interpolator = RectBivariateSpline(
             x, y, self._data.T, kx=degx, ky=degy, s=smoothness)
 
@@ -677,9 +671,9 @@ class EPSFModel(FittableImageModel):
             # find indices of pixels that are outside the input pixel
             # grid and set these pixels to the 'fill_value':
             invalid = (((xi < 0) | (xi > (self._nx - 1)
-                                    / self.oversampling[0])) |
+                                    / self.oversampling[1])) |
                        ((yi < 0) | (yi > (self._ny - 1)
-                                    / self.oversampling[1])))
+                                    / self.oversampling[0])))
             evaluated_model[invalid] = self._fill_value
 
         return evaluated_model